requirements.txt
psycopg2-binary
sqlalchemy
psutil
sqlfluff
matplotlib
pandas
click
pyyaml
rich>=13.0
